Rename index to class names
